# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Siga o seu candidato API",
    "description": "\nAPI pública e anônima para pesquisar quem governa o país, a partir de bases\nde dados abertas do governo. Rotas em `/api/v1/admin` exigem login de\nadministrador.\n",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Liveness probe",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Healthz Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Readiness probe (database and Redis)",
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Readyz Readyz Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/senators": {
      "get": {
        "tags": [
          "senators"
        ],
        "summary": "Buscar senadores em exercício",
        "operationId": "search_senators_api_v1_senators_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Nome (sem acento ok)",
              "title": "Q"
            },
            "description": "Nome (sem acento ok)"
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Za-z]{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sigla da UF, ex.: SP",
              "title": "Uf"
            },
            "description": "Sigla da UF, ex.: SP"
          },
          {
            "name": "party",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sigla do partido, ex.: PT",
              "title": "Party"
            },
            "description": "Sigla do partido, ex.: PT"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParliamentarianPage"
                }
              }
            }
          },
          "503": {
            "description": "Fonte indisponível e sem dados armazenados"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/senators/{senator_id}": {
      "get": {
        "tags": [
          "senators"
        ],
        "summary": "Detalhar senador",
        "operationId": "get_senator_api_v1_senators__senator_id__get",
        "parameters": [
          {
            "name": "senator_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Senator Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParliamentarianOut"
                }
              }
            }
          },
          "404": {
            "description": "Parlamentar não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/state-deputies": {
      "get": {
        "tags": [
          "state-deputies"
        ],
        "summary": "Buscar deputados estaduais e distritais eleitos",
        "operationId": "search_state_deputies_api_v1_state_deputies_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Nome (sem acento ok)",
              "title": "Q"
            },
            "description": "Nome (sem acento ok)"
          },
          {
            "name": "uf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Za-z]{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sigla da UF, ex.: SP",
              "title": "Uf"
            },
            "description": "Sigla da UF, ex.: SP"
          },
          {
            "name": "party",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sigla do partido, ex.: PT",
              "title": "Party"
            },
            "description": "Sigla do partido, ex.: PT"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParliamentarianPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/state-deputies/{deputy_id}": {
      "get": {
        "tags": [
          "state-deputies"
        ],
        "summary": "Detalhar deputado estadual",
        "operationId": "get_state_deputy_api_v1_state_deputies__deputy_id__get",
        "parameters": [
          {
            "name": "deputy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deputy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParliamentarianOut"
                }
              }
            }
          },
          "404": {
            "description": "Parlamentar não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ParliamentarianOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "party": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party"
          },
          "uf": {
            "type": "string",
            "title": "Uf"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "mandate_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mandate Start"
          },
          "mandate_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mandate End"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "full_name",
          "role",
          "party",
          "uf",
          "status",
          "email",
          "photo_url",
          "profile_url",
          "mandate_start",
          "mandate_end",
          "source",
          "fetched_at"
        ],
        "title": "ParliamentarianOut"
      },
      "ParliamentarianPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ParliamentarianOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ParliamentarianPage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}